Integrate functions for hash parameters.
